backend/requirements.txt
openai==0.28.0
pydantic>=2.8.2
databases>=0.9.0
supabase>=2.6.0
fastapi>=0.112.0
psycopg2-binary>=2.9.9
python_dotenv>=1.0.1
requests>=2.32.3
uvicorn>=0.30.5